+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - Feature(angular-router): Add librarySaNavigateComponent
 - Bugfix(core): undefined is filtered from selections, so TS needed to reflect that.
 - Bugfix(core): Fix buildAdapter type inference
+- Refactor(core)!: Remove deprecated buildSelectors function
 
 ## 0.59.0 2022-11-09